Added coverage and split VireoMerged.cpp
