+++ source/include/TypeAndDataManager.h
@@ -28,6 +28,7 @@ SDG
 
 #define STL_MAP
 
+#include <cmath>
 #include <stdlib.h> // abs()
 #include <new>      // for new placement
 
@@ -328,9 +329,23 @@ IntMax ReadIntFromMemory(TypeRef type, void* pData);
 NIError WriteIntToMemory(TypeRef type, void* pData, IntMax value);
 Double ReadDoubleFromMemory(TypeRef type, void* pData);
 NIError WriteDoubleToMemory(TypeRef type, void* pData, Double value);
-Double RoundToEven(Double value);
-Single RoundToEven(Single value);
 IntMax ConvertNumericRange(EncodingEnum encoding, Int32 size, IntMax input);
+//------------------------------------------------------------
+//! Banker's rounding for Doubles.
+inline Double RoundToEven(Double value)
+{
+    return rint(value);
+}
+//------------------------------------------------------------
+//! Banker's rounding for Singles.
+inline EMSCRIPTEN_NOOPT Single RoundToEven(Single value)
+{
+#if kVireoOS_emscripten
+    return rint((Double)value);
+#else
+    return rintf(value);
+#endif
+}
 
 //------------------------------------------------------------
 //! Stack based class to manage a threads active TypeManager.
